+++ src/conversion/utils/titleToName.test.ts
@@ -15,11 +15,16 @@ describe('how titleToName works', () => {
 
     it('should make name url', () => {
         expect(titleToName(`https://promptbook.studio/webgpt/write-website-content.ptbk.md`)).toBe(
-            `https://promptbook.studio/webgpt/write-website-content.ptbk.md`,
+            `promptbook-studio-webgpt-write-website-content-ptbk-md`,
+        );
+
+        expect(titleToName(`https://promptbook.studio/webgpt/write-website-content.html`)).toBe(
+            `promptbook-studio-webgpt-write-website-content`,
         );
     });
 
     it('should make relative file path', () => {
-        expect(titleToName(`../webgpt/write-website-content.ptbk.md`)).toBe(`../webgpt/write-website-content.ptbk.md`);
+        expect(titleToName(`../webgpt/write-website-content.ptbk.md`)).toBe(`webgpt-write-website-content-ptbk-md`);
+        expect(titleToName(`../webgpt/write-website-content.ptbk.html`)).toBe(`webgpt-write-website-content-ptbk-html`);
     });
 });