+++ CHANGELOG.md
@@ -526,14 +526,16 @@ Knowledge scrapers [🐝]
 -   Allow to import `.docx` files with knowledge `.docx` -(Pandoc)-> `.md` <!-- <- TODO: [🐝] !!!!!! Implement -->
 -   Allow to import `.doc` files with knowledge `.doc` -(LibreOffice)-> `.docx` -(Pandoc)-> `.md` <!-- <- TODO: [🐝] !!!!!! Implement -->
 -   Allow to import `.rtf` files with knowledge `.rtf` -(LibreOffice)-> `.docx` -(Pandoc)-> `.md` <!-- <- TODO: [🐝] !!!!!! Implement -->
+-   Allow to import websites with knowledge
 -   Add new error `KnowledgeScrapeError`
 -   Add `FilesystemTools`
+-   Changed behavior of `titleToName` when passing URLs or file paths
 
 ### `0.72.0` _(2024-09-)_
 
 Support for [Assistants API (GPTs)](https://platform.openai.com/docs/assistants/overview) from OpenAI
 
-- `getClient` methods are public
+-   `getClient` methods are public
 
 ## Drafts